contract updates

+++ contracts/simple.sol
@@ -1,12 +1,13 @@
 pragma solidity ^0.5.0;
 
 contract Simple {
-  int foo;
-  constructor(int f) public payable
-  {
+  int256 foo;
+
+  constructor(int256 f) public payable {
     foo = f;
   }
-  function set(int f) public payable {
+
+  function set(int256 f) public payable {
     foo = f;
   }
 }
