+++ contracts/tester.sol
@@ -1,16 +1,17 @@
 pragma solidity ^0.5.0;
 
 contract Tester {
-  int foo;
-  constructor(int f) public payable
-  {
+  int256 foo;
+
+  constructor(int256 f) public payable {
     foo = f;
   }
-  function incPayable(int i) public payable {
+
+  function incPayable(int256 i) public payable {
     foo += i;
   }
 
-  function inc(int i) public {
+  function inc(int256 i) public {
     foo += i;
   }
 
@@ -22,11 +23,12 @@ contract Tester {
     }
   }
 
-  function getAndIncFoo() public returns (int) {
+  function getAndIncFoo() public returns (int256) {
     foo += 1;
     return foo;
   }
-  function getFoo() public view returns (int) {
+
+  function getFoo() public view returns (int256) {
     return foo;
   }
 }
